+++ backend/Main/Main/Queries/DTO/PostInfoDTO.cs
@@ -8,8 +8,8 @@
         public string Headline { get; set; }
         public string Body { get; set; }
         public string SourceName { get; set; }
-        public int SRR { get; set; }
-        public int PTS { get; set; }
+        public double SRR { get; set; }
+        public double PTS { get; set; }
         public DateTime DateTime { get; set; }
         public string IsAdmin { get; set; }
         public DateTime TimeCreated { get; set; }
